+++ packages/drizzle/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @authhero/drizzle
 
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.47.0
+
 ## 0.2.6
 
 ### Patch Changes